azure-requirements.txt
# Core ML/AI Libraries (CPU versions for Azure)
torch==2.3.0+cpu
torchvision==0.18.0+cpu
torchaudio==2.3.0+cpu
sentence-transformers
faiss-cpu

# LangChain and NVIDIA AI
langchain
langchain-nvidia-ai-endpoints

# PDF Processing
pdfplumber

# Data Processing
numpy>=1.21
pandas
scikit-image==0.18.0

# Image Processing (for document analysis)
opencv-python
Pillow
tifffile==2020.9.3
imageio

# Web Framework (for API)
fastapi
uvicorn

# HTTP Requests
requests

# Environment Management
python-dotenv

# Utilities
tqdm
matplotlib
tensorboard
tensorboardX
ipython
dominate==2.6.0
lpips==0.1.3

# Optional: Keras compatibility fix
tf-keras

# Azure specific
wfastcgi 